Updates to get integrated GWC working.
This pulls in the latest from the GeoServer wfs2-gss branch to get a fix for cache truncation on edits.  As the cache is being truncated, the client is getting a lot of incomplete or invalid responses (haven't checked which yet).  To work around this, we request images 3 times before giving up.