Feature editor height management (closes #131).
In cases where the map portal height is shorter than the attribute form, we avoid making the window too large.  With each change in child panel visibility, the height is updated to fit the contents if possible.  When the user changes the window height, we keep track of this as the preferred height.